edit speed

+++ AvoidMaster/Sprite/ObstacleMangager.cs
@@ -15,7 +15,7 @@ namespace AvoidMaster.Sprite
         public Rectangle GameBoundaries { get; }
         public GameObjects GameObjects { get; }
         private double timeSinceLastObastacle;
-        public float SpeedCreateObstacle { get; set; }
+        public double SpeedCreateObstacle { get; set; }
 
         private GraphicsDevice graphicsDevice;
         public ObstacleMangager(Rectangle gameBoundaries, GraphicsDevice graphicsDevice, GameObjects gameObjects)
@@ -23,7 +23,7 @@ namespace AvoidMaster.Sprite
             GameBoundaries = gameBoundaries;
             this.graphicsDevice = graphicsDevice;
             GameObjects = gameObjects;
-            SpeedCreateObstacle = 0.5f;
+            SpeedCreateObstacle = 0.5;
             CreateEnemy();
         }
